% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_e_other.R
\name{ro_e_detect_axis_orientation}
\alias{ro_e_detect_axis_orientation}
\title{Detects chart orientation from the echarts4r object.}
\usage{
ro_e_detect_axis_orientation(e)
}
\arguments{
\item{e}{An echarts4r object.}
}
\value{
One of "horizontal" or "vertical".
}
\description{
Returns "horizontal" for column/line charts (categories on xAxis) and
"vertical" for bar charts (categories on yAxis after e_flip_coords()).
}
\details{
Detection strategy: after e_flip_coords(), echarts4r moves category data to
yAxis. Check whether e$x$opts$yAxis has a non-NULL \code{data} element.
}
\seealso{
Other echarts4r:
\code{\link[=ro_e_build_keyboard_nav_js]{ro_e_build_keyboard_nav_js()}},
\code{\link[=ro_e_build_onrender_js]{ro_e_build_onrender_js()}},
\code{\link[=ro_e_keyboard_nav]{ro_e_keyboard_nav()}},
\code{\link[=ro_e_keyboard_nav_axis_js]{ro_e_keyboard_nav_axis_js()}},
\code{\link[=ro_e_keyboard_nav_legend_js]{ro_e_keyboard_nav_legend_js()}},
\code{\link[=ro_e_theme]{ro_e_theme()}},
\code{\link[=ro_e_theme_list]{ro_e_theme_list()}}
}
\concept{echarts4r}
\keyword{internal}
