% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.echarts-package.R
\docType{package}
\name{ROvis.echarts-package}
\alias{ROvis.echarts}
\alias{ROvis.echarts-package}
\title{ROvis.echarts: Accessible Echarts With Rijksoverheid Styling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ROvis.echarts is an R package that provides a comprehensive suite of functions for making echart plots in rijksoverheid style.
}
\author{
\strong{Maintainer}: ROvis team \email{spin@rivm.nl}

Authors:
\itemize{
  \item ROvis team \email{spin@rivm.nl}
}

}
\keyword{internal}
